chore(gradle): Add support for local dev on M1 processors
Problem

Our tests require that we run the `mysql:5.7` docker container image, and this image doesn't have ARM binaries. This leads to errors when trying to run locally on the ARM-based X1s.

We spin up containers two different ways:

1. In oss/keel-sql/keel-sql.gradle, there's a `liquibaseUpdate` section that invokes docker to spin up a mysql image.
2. In our tests that involve the database, we use the testcontainers library to spin up MySQL 5.7 images as needed.

Solution

For the keel-sql.gradle case, we now have a `--platform=linux/amd64` flag that tells Docker to use the x86 binaries, which run fine with our Macs (assuming you have [Rosetta](https://support.apple.com/en-us/HT211861) installed).

For the testcontainers case, the latest version of testcontainers defaults to linux/amd64 images if it encounters an error, so this PR bumps to the latest version (1.16.2).

Note that the current version of kork-sql that we're locked to was locking to the old version of testcontainers, so I had to do a `generateLock saveLock`, which bumped all of the libraries.

Squashed commit of the following:

commit 9854fe05391c0b5fdccfacb120daf48dc69091ad
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Mon Nov 15 15:54:12 2021 -0800

    chore(gradle): bump dependency locks

    Bump the dependency locks so we can actually get the latest version of
    testcontainers.

commit 9c946172fe06ec7a09f61895162236fb993ae0b1
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Mon Nov 15 15:40:53 2021 -0800

    chore(gradle): run mysql:5.7 using arm

commit cfac897e8fe9638a4ec049407d7efb216b9468e7
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Mon Nov 15 15:38:36 2021 -0800

    chore(gradle): bump testcontainers for running x86 on M1

    The latest version of testcontainers should fall back to x86 binaries if
    ARM binaries are not available for docker images like mysql:5.7.

(cherry picked from commit 83556b88e6f71b1b71b833589acf39f5f73e5969)