fix(graphql): change variant for preview environment
Problem

If a delivery config contains a graphql/schema resource, then it will fail to create a preview environment, because GraphqlSchemaSpec doesn't define deepRename.
We need to decide how to treat graphql/schema resources in preview environments. We cannot add a suffix to them the way we do with other resources, because these resources don't have a name in the same sense that other resources do.

See MD-24

Implemented Solution: change variant for preview env
In this solution, in the preview environment, we change whcih variant we publish to.
We introduce a new field to the graphql/schema resource called "previewVariant". This will be the variant that MD will publish to for the graphql/schema resource in the preview environment.

For example:
resources:
- kind: graphql/schema@v1
  spec:
    application: myapp
    artifactReference: myapp-docker
    edge: enterprise
    registryEnv: test
    previewVariant: integrationc

When this resource is copied into a preview environment, the variant gets set to integrationc.

Default values

Because there is a fixed set of variants, I had to choose a default. The valid values are
staging, integration, integrationa, integrationb, integrationc, integrationd, integratione
I chose integratione because I assumed that would be the one least likely to have a collision, but that was just an arbitrary decision.

Issues

Only one preview enviornment at a time
With this scheme, we can only support one preview environment at a time. Otherwise, MD will fail with an error because it will try to create two identical graphql/schema resources.
A better solution would be to dynamically generate variants. However, due to how the registry is implemented, this is not yet possible, we only have a fixed set of valid variants to work with.
I have an action item to follow up with the BFG team about this.

Squashed commit of the following:

commit 34d9eb75699b571473840f663e48d32f7f5213a5
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Thu Dec 2 13:48:38 2021 -0800

    fix(pr): cosmetic change to test

commit b4735bf4ecf8a04028ff0836def2cabe8f78abcd
Merge: f2bd6a4ee 4962ae380
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Thu Dec 2 13:18:49 2021 -0800

    Merge branch 'master' into MD-24

commit f2bd6a4ee37eef5827f0e6799bb06d3fcfc1ba1a
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Thu Dec 2 13:18:22 2021 -0800

    fix(pr): remove accidental import

commit d338afa56bf0737c23f3fd3c1b78478609e5a0df
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Wed Nov 24 14:01:41 2021 -0800

    Problem

    If a delivery config contains a graphql/schema resource, then it will fail to create a preview environment, because GraphqlSchemaSpec doesn't define deepRename.

    We need to decide how to treat graphql/schema resources in preview environments. We cannot add a suffix to them the way we do with other resources, because these resources don't have a name in the same sense that other resources do.

    See MD-24

    Implemented Solution

    We introduce a new field to the graphql/schema resource called "previewVariant". This will be the variant that MD will publish to for the graphql/schema resource in the preview environment.

    For example:

    ```
    resources:
    - kind: graphql/schema@v1
      spec:
        application: myapp
        artifactReference: myapp-docker
        edge: enterprise
        registryEnv: test
    		previewVariant: integrationc
    ```

    Some issues:

    Only one preview enviornment at a time

    With this scheme, we can only support one preview environment at a time.  Otherwise, MD will fail with an error because it will try to create two identical graphql/schema resources.

    Default values

    Variant names are fixed, and so I had to choose a default. The valid values are
    staging, integration, integrationa, integrationb, integrationc, integrationd, integratione

    I chose `integratione` because I assumed that would be the one least likely to have a collision, but that was just an arbitrary decision.

(cherry picked from commit b94d38bdfba480bf33d4f93a4bdc9e43ad19778e)