fix(ec2): fix step scaling policy issue
There were two issues with step scaling policies on EC2:

We weren't passing the serverGoupName field in the EC2 case (we were in the Titus case), which cause the orca upsert scaling policy task to fail (see MD-48).

The metricIntervalLowerBound and metricIntervalUpperBound fields that are returned by clouddriver are lower-case for EC2, where they are upper-case for Titus.

We confirmed with John Christopher Smalley that we can send either uppercase or lowercase, so I modified the code to send them as lowercase, and used a JsonAlias field to handle the Titus case where the clouddriver response is upper case.

Fixes MD-48

Squashed commit of the following:

commit 7d7d6f5fd5e07bb393dfc143900e5bd65dee153c
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Tue Jan 4 12:54:43 2022 -0800

    Add json aliases

commit 638fa6566dec4340022ffc1e0b874160e306aa04
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Tue Jan 4 12:53:06 2022 -0800

    s/MetricIntervalUpperBound/metricIntervalUpperBound

commit fd177fedf9534ccea7175e8390acb78c96a375b1
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Tue Jan 4 12:52:19 2022 -0800

    s/MetricIntervalLowerBound/metricIntervalLowerBound

commit 0a0c8086bd23f9627f3c918e6ed859f722561f29
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Tue Jan 4 12:50:23 2022 -0800

    fix(ec2): send serverGroupName when setting scaling policy

(cherry picked from commit 6889cbf64d7c1d33b81ee35ee2ff7706a1b72934)