fix(titus): send platform sidecar payload
Allow the TitusClusterServerGroupSpec.platformSidecars field to be null so that we handle the case where override objects are present in the TitusClusterSpec.

Problem

Whenever there is an override in a Titus cluster spec, we would resolve platform sidecar list to empty, even if there were sidecars.

That's because we were assuming the override value would be null if not present, where it was actually an empty list.

In practice, there's always an override added by the ContainerAttributesResolver, but the unit tests missed it because we don't apply the resolveer in unit tests, so tests passed but it didn't work.

Solution

Allow the TitusClusterServerGroupSpec.platformSidecars field to be null so that we handle the case where override objects are present in the TitusClusterSpec.

Note that we still do not allow the resolved TitusServerGroup.platformSidecars field to be null, because clouddriver will return an empty list, and so we always want this to resolve to an empty list.

Fixes TITUS-5818

Squashed commit of the following:

commit 150c9a5a31285ac5a97b1eb48ec1b49a466d4458
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Wed Feb 9 14:44:05 2022 -0800

    fix(titus): send platform sidecar payload

    Problem

    Whenever there is an override in a Titus cluster spec, we would resolve platform sidecar list to empty, even if there were sidecars.
    That's because we were assuming the override value would be null if not present, where it was actually an empty list.

    In practice, there's always an override added by the ContainerAttributesResolver, but the unit tests missed it because we don't apply the resolveer in unit tests, so tests passed but it didn't work.

    Solution

    ALlow the TitusClusterServerGroupSpec.platformSidecars field to be null so that we handle the case where override objects are present in the TitusClusterSpec.

    Note that we still do not allow the resolved TitusServerGroup.platformSidecars field to be null, because clouddriver will return an empty list, and so we always want this to resolve to an empty list.

    Fixes TITUS-5818

(cherry picked from commit 1ad3189fb736162653b3d035212dfacdd54d360b)