fix(test): explicitly set spring.application.name
Problem

The ArtifactControllerTests test class, when run on a laptop, fails with an exception.
This is because the Spring setup for this test will instantiate a LaptopTaskQueueNamer bean, and that bean queries the environment for the spring.application.name field.

This test is configured to use a MockEnvironment, so even though we have spring.application.name defined in application.properties in the oss/keel-web/src/test/resources directory, this test does not actually use that file to construct an environment.

Solution

Explicitly set the spring.application.name property on the MockEnvironment object.

Squashed commit of the following:

commit 714e7f551648624db0d453bd42de0eceb731ad87
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Mon Feb 28 21:15:40 2022 -0800

    fix(test): explicitly set spring.application.name

    Problem

    The ArtifactControllerTests test class, when run on a laptop, fails with
    an exception.

    This is because the Spring setup for this test will instantiate a
    LaptopTaskQueueNamer bean, and that bean queries the environment for the
    "spring.application.name" field.

    This test is configured to use a MockEnvironment, so even though we have
    spring.application.name defined in application.properties in the
    oss/keel-web/src/test/resources directory, this test does not actually
    use that file to construct an environment.

    Solution

    Explicitly set the spring.application.name property on the
    MockEnvironment object.

(cherry picked from commit 11669e7ce57238de1aa2fb47120302e128846828)