fix(scheduler): toggles around waiting for batch
Add feature toggles so that we can turn off waiting for all of the resource checks in a batch to complete.

Problem

keel is processing too slowly, which is causing issues for users.
We don't have a solid understanding of where the time is going.
One issue I noticed is that, in each of the scheduler loops, we block
waiting for all of the coroutines in the batch to complete before
continuining. This means that each scheduler function can only proceed
as quickly as the slowest coroutine in the batch.
According to atlas, processing a batch takes about 3.4 seconds, but a single resource takes about 1.7 second, and so keel has to wait for the slower resource checks in the batch to complete before it can continue to make progress.

Solution

I created a set of feature toggles so that we can turn off waiting for all of the resource checks in a batch to complete. When this feature toggle is off, the checks will all be completely asynchronous: we won't wait for previous ones to complete anymore.
The default setting is the current behavior, so the initial rollout should not change behavior.

Notes

This is a pretty big behavioral change for keel. We're going to switch to asynchronously launching a batch of resource checks every second, and
not waiting for it to complete. I expect us to process through the
resources a lot more quickly.

How I'm going to test this

As mentioned above, the default value of the featue toggles is to keep the current behavior (block waiting for the batch to complete).

I'm going to start by selectively turning off the wait in prestaging for resource checks. If that looks OK, then I'll test it out in main.

I'll be watching the dashboard for resource contention/exhaustion, especially the database and CPU/memory. Database because we'll have a lot more parallelism potentially, and CPU/memory because we may end up creating a lot of coroutines, and I'm not sure what the system impact is going to be.

Squashed commit of the following:

commit 72975e95520f2d18a41ed3db8ab22d357b68a23e
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Mon Apr 18 14:40:10 2022 -0700

    fix(scheduler): toggles around waiting for batch

    Problem

    keel is processing too slowly, which is causing issues for users.
    We don't have a solid understanding of where the time is going.

    One issue I noticed is that, in each of the scheduler loops, we block
    waiting for all of the coroutines in the batch to complete before
    continuining. This means that each scheduler function can only proceed
    as quickly as the slowest coroutine in the batch.

    Solution

    I created a set of feature toggles to enable us to turn this on/off.

    The default setting is the current behavior.

    Expected behavior

    We're going to switch to asynchronously launching work every second, and
    not waiting for it to complete. I expect us to process through the
    resources more quickly.

    TOday we're averaging 1.7 s to complete a resource check, and we're going to be
    launching 5 new checks every second. This could end up launching a lot
    of coroutines, or maybe leading to database contention.

(cherry picked from commit 888e8746713bd6bc50042e93745664ae476f598d)