fix(clouddriver): pass vpc when looking up security groups
Problem

When keel looks up security group id in clouddriver, it doesn't pass the
vpc id. This causes problems if a security group with the same
name/region/account is associated with a different VPC. This led to
situations where we got the security group for nf-infrastructure associated with vpc1 instead of vpc0.

Solution

Always pass the vpc id when doing security group lookups to ensure we get the right one.

Note

For Titus server groups, keel doesn't store the vpc, only the subnets.
Because this bug is currently only affecting us in EC2 apps, right now
we are falling back to the current behavior for Titus apps by passing
"*" as the VPC, which is the behavior that clouddriver uses when no vpc
is passed in the security group query.

Fixes MD-189

Squashed commit of the following:

commit fc9aea5417c547b3d139e5ce0021e44a28302c5e
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Thu May 12 15:04:45 2022 -0700

    fix(clouddriver): pass vpc when looking up security groups

    Problem

    When keel looks up security group id in clouddriver, it doesn't pass the
    vpc id. This causes problems if a security group with the same
    name/region/account is associated with a different VPC. This led to
    situations where we got the wrong security group.

    Solution

    Always pass the vpc id when doing security group lookups.

    Note

    For Titus server groups, keel doesn't store the vpc, only the subnets.
    Because this bug is currently only affecting us in EC2 apps, right now
    we are falling back to the current behavior for Titus apps by passing
    "*" as the VPC, which is the behavior that clouddriver uses when no vpc
    is passed in the security group query.

(cherry picked from commit 4d4c156d32883fed7796b822d2d4751abe19d5f3)