fix(titus): add vpc to Location
Problem
-------

We are getting naming collisions with security groups on vpcs other than
vpc0, for Titus apps. Because we don't explicitly pass the vpc to
clouddriver when we resolve security group ids, we sometimes have the
wrong security group id when trying to do a deploy, which leads to a
failure.

This happened with the gamestats app, where we resolved the
`nf-datacenter` security group to one in `vpc1_cosmos` instead of `vpc0`. See
<https://netflix.slack.com/archives/C0DSWME1E/p1652810794738549>

Solution
--------

Always pass the vpcId when resolving security group ids. This requires
us to plumb vpc info all of the way through the system.

Changes of interest
-------------------

The `TitusServerGroup.Location` data class now has a `vpc` field that
encodes the vpc name. It defaults to the default VPC so this shouldn't
break any existing stuff.

The `SimpleLocations.vpc` field is now required instead of optional.

Fixes MD-205.

Squashed commit of the following:

commit 2ddba57819ed28abe8b1c762e635ee1fb509885e
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Fri May 20 16:00:31 2022 -0700

    add comment

commit 0ecf6849d59f227c4db57f735441e0b8e1606214
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Wed May 18 13:42:31 2022 -0700

    fix(titus): add vpc to Location

    Problem
    -------

    We are getting naming collisions with security groups on vpcs other than
    vpc0, for Titus apps. Because we don't explicitly pass the vpc to
    clouddriver when we resolve security group ids, we sometimes have the
    wrong security group id when trying to do a deploy, which leads to a
    failure.

    This happened with the gamestats app, where we resolved the
    `nf-datacenter` security group to one in `vpc1_cosmos` instead of `vpc0`. See
    <https://netflix.slack.com/archives/C0DSWME1E/p1652810794738549>

    Solution
    --------

    Always pass the vpcId when resolving security group ids. This requires
    us to plumb vpc info all of the way through the system.

    Changes of interest
    -------------------

    The `TitusServerGroup.Location` data class now has a `vpc` field that
    encodes the vpc name. It defaults to the default VPC so this shouldn't
    break any existing stuff.

    The `SimpleLocations.vpc` field is now required instead of optional.

(cherry picked from commit 9cd35a4c43ee85ddea11a8b2adf3268a1c21e907)