fix(titus): use correct account for network lookup
When keel cloned an existing Titus server group (e.g., for re-deploy), we were passing the wrong value for the account when looking up the VPC id in clouddriver.

For example, we were passing the account name as "titustestvpc" where the call exepcted the account name "test".

This resulted in the following error:

    Request Failed com.netflix.spinnaker.keel.clouddriver.ResourceNotFound: VPC network named vpc0 not found in titustestvpc:us-east-1

The fix is to call getAwsAccountNameForTitusAccount to get the awsAccount. We were already doing this in the deploy call path, but not in the redeploy (clone) call path.

Fixes MD-248

Squashed commit of the following:

commit a748780ffa89d354a18504cf9e3cdf071078e6b6
Author: Lorin Hochstein <lhochstein@netflix.com>
Date:   Wed May 25 13:44:12 2022 -0700

    fix(titus): use correct account for network lookup

    When keel cloned an existing Titus server group (e.g., for re-deploy), we were
    passing the wrong value for the account when looking up the VPC id in
    clouddriver.

    For example, we were passing the account name as "titustestvpc" where the call exepcted the account name "test".

    This resulted in the following error:

      Request Failed com.netflix.spinnaker.keel.clouddriver.ResourceNotFound: VPC network named vpc0 not found in titustestvpc:us-east-1

    The fix is to call getAwsAccountNameForTitusAccount to get the
    awsAccount. We were already doing this in the deploy call path, but not
    in the redeploy (clone) call path.

    Fixes MD-248

(cherry picked from commit 9930c739a9095daf0c46a3a901b20e3972f36f03)