Add new lag metrics
Add "lag metrics" which contain a duration for maximum of the last time
that a resource was checked in a control loop iteration.

Add this metric for the control loops such as resource, verification,
environment, etc.

Removed the "AboutToBeChecked" event since it is now redundant.

New resources or resources where a recheck was triggered are not
considered when calculating the metric since there's not really a "last
checked time" for these cases.